requirements.txt
Django==4.2
celery==5.3.1
redis==4.5.5
django-celery-results==2.5.1
djangorestframework==3.16.0
django-filter==25.1
django-redis==5.4.0
flower==2.0.1
qrcode==8.1
django-celery-email==3.0.0
requests==2.32.3
pillow==11.1.0
psycopg2-binary>=2.9.3
djoser==2.3.1
